Update functions and aliases
Add .git_aliases
Better colors
*dev-functions have better venv handling

+++ .bashrc
@@ -1,3 +1,12 @@
+# Better colors than standard ubuntu
+if [ "$color_prompt" = yes ]; then
+    # PS1='${debian_chroot:+($debian_chroot)}\[\033[01;32m\]\u@\h\[\033[00m\]:\[\033[01;34m\]\w\[\033[00m\]\$ '
+    PS1='${debian_chroot:+($debian_chroot)}\[\033[01;32m\]\u@\h\[\033[00m\]:\[\033[01;32m\]\w\[\033[00m\]\$ '
+else
+    PS1='${debian_chroot:+($debian_chroot)}\u@\h:\w\$ '
+fi
+LS_COLORS=$LS_COLORS:'di=0;32:' ; export LS_COLORS
+
 # Make dir and cd into it
 mkcd ()
 {
@@ -16,8 +25,16 @@ cl ()
   fi
 }
 
-# Select development environment, with tab completion
+# Select development environment
+# Enable tab completion in .dev-completion.bash
 dev ()
+{
+    source ~/code/venvs/"$1"/bin/activate &&
+    cd -P -- ~/code/workspaces/"$1"
+}
+# Select development environment and python venv
+# Enable tab completion in .dev-completion.bash
+devv ()
 {
     source ~/code/venvs/"$1"/bin/activate &&
     cd -P -- ~/code/workspaces/"$1"
@@ -27,8 +44,6 @@ dev ()
 mkdev ()
 {
     if [ "$1" ]; then
-        python3 -m venv ~/code/venvs/"$1" &&
-        source ~/code/venvs/"$1"/bin/activate &&
         mkdir ~/code/workspaces/"$1" &&
         cd -P -- ~/code/workspaces/"$1"
     else
@@ -36,30 +51,27 @@ mkdev ()
     fi
 }
 
-# Create development environment, checkout things and start vscode
-mkdevf ()
+# Create development environment and python venv
+mkdevv ()
 {
     if [ "$1" ]; then
-        python3 -m venv ~/code/venvs/"$1"
-        source ~/code/venvs/"$1"/bin/activate
-        mkdir ~/code/workspaces/"$1"
-        cd -P -- ~/code/workspaces/"$1"
-
-        git clone "<add git clone address here>"
-        # pip install wheel tox # Install specific python packets
-        # pip install -e <cloned folder>/ # Install the cloned repo contents in dev mode
-        code . # Start vscode
+        python3 -m venv ~/code/venvs/"$1" &&
+        source ~/code/venvs/"$1"/bin/activate &&
+        mkdev "$1"
     else
         echo "Empty name supplied..."
     fi
 }
 
 # Remove development environment
+# Enable tab completion in .dev-completion.bash
 rmdev ()
 {
     cd ~/code/workspaces
-    deactivate
-    rm -rf ~/code/venvs/"$1"
+    if [ -d ~/code/venvs/"$1" ]; then
+        deactivate &&
+        rm -rf ~/code/venvs/"$1"
+    fi
     rm -rf ~/code/workspaces/"$1"
 }
 
@@ -69,4 +81,17 @@ source ~/.dev-completion.bash
 # set XON/XOFF active
 [[ $- == *i* ]] && stty -ixon
 
+
+# Fix password entry for gpg signing over ssh
+# sudo update-alternatives --config pinentry
+#   select pinentry-curses
+# gpg-connect-agent reloadagent /bye
+export GPG_TTY=$(tty)
+
+# git aliases
+source ~/.git_aliases
+gh ()
+{
+    cat ~/.git_aliases
+}
